+++ src/Symfony/Component/Runtime/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 ---
 
  * The component is not experimental anymore
+ * Add options "env_var_names" to `GenericRuntime` and `SymfonyRuntime`
 
 5.3.0
 -----